+++ src/Application/UseCases/Appointment/UpdateAppointment/UpdateAppointmentCommand.cs
@@ -1,3 +1,3 @@
 namespace Application.UseCases.Appointment.UpdateAppointment;
 
-public sealed record UpdateAppointmentCommand(Guid Id, string Description, DateTime KeepDate, byte AmountHours, Guid AssignmentId, Guid UserId) : BaseCommand, IRequest<OperationResult>;
+public sealed record UpdateAppointmentCommand(Guid Id, string Description, DateTime KeepDate, int AmountHours, Guid AssignmentId, Guid UserId) : BaseCommand, IRequest<OperationResult>;
