+++ src/Application/UseCases/Assignment/UpdateAssignment/UpdateAssignmentCommand.cs
@@ -1,3 +1,3 @@
 namespace Application.UseCases.Assignment.UpdateAssignment;
 
-public sealed record UpdateAssignmentCommand(Guid Id, string Name, string Description, DateTime StartDate, DateTime EndDate, byte AmountHours, Guid ProjectId, Guid WorkflowId, Guid UserId, Guid AssignmentTypeId) : BaseCommand, IRequest<OperationResult>;
+public sealed record UpdateAssignmentCommand(Guid Id, string Name, string Description, DateTime StartDate, DateTime EndDate, int AmountHours, Guid ProjectId, Guid WorkflowId, Guid UserId, Guid AssignmentTypeId) : BaseCommand, IRequest<OperationResult>;
