+++ src/Application/UseCases/Workflow/CreateWorkflow/CreateWorkflowCommand.cs
@@ -1,3 +1,3 @@
 namespace Application.UseCases.Workflow.CreateWorkflow;
 
-public sealed record CreateWorkflowCommand(string Name, byte Order, Guid Id = default) : BaseCommand, IRequest<OperationResult>;
+public sealed record CreateWorkflowCommand(string Name, int Order, Guid Id = default) : BaseCommand, IRequest<OperationResult>;
