+++ src/Application/UseCases/Workflow/UpdateWorkflow/UpdateWorkflowCommand.cs
@@ -1,3 +1,3 @@
 namespace Application.UseCases.Workflow.UpdateWorkflow;
 
-public sealed record UpdateWorkflowCommand(Guid Id, string Name, byte Order) : BaseCommand, IRequest<OperationResult>;
+public sealed record UpdateWorkflowCommand(Guid Id, string Name, int Order) : BaseCommand, IRequest<OperationResult>;
