+++ src/Application/UseCases/Workflow/UpdateWorkflow/UpdateWorkflowCommandValidator.cs
@@ -11,6 +11,6 @@ public sealed class UpdateWorkflowCommandValidator : AbstractValidator<UpdateWor
             .NotEmpty().WithMessage("Name is required.");
 
         RuleFor(x => x.Order)
-            .GreaterThan((byte)0).WithMessage("Order must be greater than 0.");
+            .GreaterThan((int)0).WithMessage("Order must be greater than 0.");
     }
 }
