+++ src/Domain/Repositories/IAssignmentRepository.cs
@@ -6,5 +6,5 @@ public interface IAssignmentRepository
     Task<Assignment?> GetAssignmentById(Guid id);
     Task<List<Assignment?>?> ListAssignments();
     Task<bool> RemoveAssignment(Guid id);
-    Task<bool> UpdateAssignment(Guid id, string name, string description, DateTime startDate, DateTime endDate, byte amountHours, Guid projectId, Guid workflowId, Guid userId, Guid assignmentTypeId);
+    Task<bool> UpdateAssignment(Guid id, string name, string description, DateTime startDate, DateTime endDate, int amountHours, Guid projectId, Guid workflowId, Guid userId, Guid assignmentTypeId);
 }
